requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-decouple==3.8
pytz==2021.1
tqdm